replace integer with long
co-276

+++ src/main/groovy/edu/oregonstate/mist/directoryapi/DirectoryEntity.groovy
@@ -18,5 +18,5 @@ class DirectoryEntity {
     String faxNumber
     Email emailAddress
     String username
-    Integer osuuid
+    Long osuuid
 }
